merged e3a0881ebc10f34c37789485abeae7a1913a056f from https://github.com/leniency/UnitySteer/
"Changed up Tether. Rather than do a hard reflect, it allows a cushion zone between an inner and outer radius. Plus, it now actually makes use of the steering weight. It applies a percentage of the steering weight to the vehicle that scales as it approaches the outer radius (past there it continues to scale above 100%). Allows for softer course adjustments. The hard reflect could be added back in at the outer radius easily enough and still leave the softer inner redirection. Seems to work nicest when the tether weight is small."